+++ CHANGELOG.md
@@ -10,8 +10,11 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 * Update syntax to be compliant with OMMBV 1.0
 * Documentation
   * Improve docstrings throughout
+* Deprecations
+  * Deprecated missions_ephem, as pyephem will no longer be updated
 * Testing
   * Add style check for docstrings
+  * Added checks for deprecation warnings
   * Improve checks in codeclimate
 
 ## [0.2.2] - 2021-06-18